# openapi.yaml
openapi: 3.1.0
info:
  title: 禱告時光 API
  description: |
    禱告手冊清單、每日內容、閱讀進度。身分驗證可二選一：(1) query 參數 email；(2) Authorization Bearer（OAuth）。建議 GPTs 使用 email 流程：先問 email → 呼叫 GET /email/books → 有訂閱則列書選書，無則提供訂閱表單連結。
    選填：任一 API 可帶 query 參數 conversation_id（同一輪對話使用相同 ID），後端會將請求寫入對話紀錄（gpt/10）並以該 ID 歸組。
  version: 1.0.0

servers:
  - url: https://prayer-time-api-43747267943.asia-northeast1.run.app
    description: Cloud Run 禱告時光 API

paths:
  /email/books:
    get:
      operationId: getEmailBooks
      summary: 依 email 查詢已訂閱書單（不需 Bearer）
      description: 依訂閱時使用的 email 查詢 Ragic 訂閱紀錄，回傳該 email 已訂閱的書單與訂閱表單連結。無需登入或 Bearer。
      parameters:
        - name: email
          in: query
          required: true
          schema: { type: string, format: email }
          description: 使用者訂閱時使用的 Email
        - name: conversation_id
          in: query
          required: false
          schema: { type: string }
          description: 選填。同一輪對話使用相同 ID，會寫入對話紀錄（gpt/10）並歸組
      responses:
        '200':
          description: 已訂閱書單與表單連結
          content:
            application/json:
              schema:
                type: object
                properties:
                  subscribed_books:
                    type: array
                    description: 該 email 已訂閱的書；每筆含 number（1 起）、book_id、book_name。使用者選數字時請用 number 對應到該筆的 book_id。
                    items:
                      type: object
                      properties:
                        number: { type: integer, description: '選項編號（1、2、3…），使用者回覆此數字時請用本筆的 book_id' }
                        book_id: { type: string }
                        book_name: { type: string }
                  subscription_form_url: { type: string, description: 尚未訂閱時可提供此連結；僅在 all_books_subscribed 為 false 時顯示' }
                  user_name: { type: string, nullable: true, description: '訂閱表填寫的姓名，取得後可用於問候' }
                  all_books_subscribed: { type: boolean, description: 'true 表示已訂閱所有可用手冊，僅顯示書單、不顯示訂閱連結；false 時可顯示「可前往訂閱」連結' }
        '400':
          description: 缺少 email 參數

  /init:
    get:
      operationId: init
      summary: 初始化（訂閱狀態 + 手冊清單）
      description: 單次呼叫回傳 hasSubscription 與 books（已訂閱才回傳 books）。身分請帶 query email 或 Authorization Bearer。
      parameters:
        - name: email
          in: query
          required: false
          schema: { type: string }
          description: 訂閱時使用的 Email（與 Bearer 二選一）
      responses:
        '200':
          description: 初始化結果
          content:
            application/json:
              schema:
                type: object
                properties:
                  hasSubscription: { type: boolean }
                  books:
                    type: array
                    items:
                      type: object
                      properties:
                        book_id: { type: string }
                        book_name: { type: string }

  /books:
    get:
      operationId: getBooks
      summary: 取得禱告手冊清單
      description: 回傳所有可選的禱告手冊（book_id、book_name），供使用者選擇。
      responses:
        '200':
          description: 手冊清單
          content:
            application/json:
              schema:
                type: object
                properties:
                  books:
                    type: array
                    items:
                      type: object
                      properties:
                        book_id: { type: string }
                        book_name: { type: string }

  /content:
    get:
      operationId: getContent
      summary: 取得指定手冊指定日的內容
      description: |
        依 book_id 與 day 回傳當日禱告內容。身分請帶 query email 或 Bearer。
        若尚未訂閱該手冊，仍回傳 200，body 含 subscribed false 與 error，請以查詢結果方式呈現。
        成功回傳內容時，後端會自動寫入一筆閱讀紀錄（gpt/7），包含 email、書、天數、閱讀時間，以及從訂閱表帶入的姓名、教會。
      parameters:
        - name: email
          in: query
          required: false
          schema: { type: string }
          description: 訂閱時使用的 Email（與 Bearer 二選一）
        - name: book_id
          in: query
          required: true
          schema: { type: string }
          description: 手冊 ID（從 /books 或 /email/books 取得）
        - name: day
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 31 }
          description: 第幾天（1–31），不傳則為「今天」對應的日
        - name: conversation_id
          in: query
          required: false
          schema: { type: string }
          description: 選填。同一輪對話使用相同 ID，會寫入對話紀錄（gpt/10）
      responses:
        '200':
          description: 當日內容或未訂閱結果（依 body 的 subscribed 判斷）
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    description: 已訂閱，回傳內容
                    properties:
                      book_id: { type: string }
                      book_name: { type: string }
                      day: { type: integer }
                      title: { type: string }
                      content: { type: string }
                  - type: object
                    description: 未訂閱
                    properties:
                      subscribed: { type: boolean, example: false }
                      error: { type: string, example: '很抱歉，您尚未訂閱禱告時光！' }
        '404':
          description: 找不到該日內容

  /titles:
    get:
      operationId: getTitles
      summary: 取得指定手冊 1–31 天標題清單
      description: 依 book_id 回傳該手冊 1–31 天的標題清單。身分請帶 query email 或 Bearer。
      parameters:
        - name: email
          in: query
          required: false
          schema: { type: string }
          description: 訂閱時使用的 Email（與 Bearer 二選一）
        - name: book_id
          in: query
          required: true
          schema: { type: string }
          description: 手冊 ID（從 /books 或 /email/books 取得）
      responses:
        '200':
          description: 標題清單或未訂閱結果
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    description: 已訂閱，回傳標題清單
                    properties:
                      book_id: { type: string }
                      book_name: { type: string }
                      titles:
                        type: array
                        description: 1–31 天的標題（無內容則為 null）
                        items:
                          type: object
                          properties:
                            day: { type: integer, minimum: 1, maximum: 31 }
                            title: { type: string, nullable: true }
                  - type: object
                    description: 未訂閱
                    properties:
                      subscribed: { type: boolean, example: false }
                      error: { type: string, example: '很抱歉，您尚未訂閱禱告時光！' }

  /progress:
    get:
      operationId: getProgress
      summary: 取得當前使用者的閱讀進度
      description: 依 email（query 或 Bearer 取得）從閱讀紀錄推導最後一次閱讀的手冊與天數。
      parameters:
        - name: email
          in: query
          required: false
          schema: { type: string }
          description: 訂閱時使用的 Email（與 Bearer 二選一）
      responses:
        '200':
          description: 閱讀進度
          content:
            application/json:
              schema:
                type: object
                properties:
                  last_book_id: { type: string, nullable: true }
                  last_book_name: { type: string, nullable: true }
                  last_day: { type: integer, nullable: true }
                  last_read_date: { type: string, nullable: true }
                  message: { type: string }

  /subscription-status:
    get:
      operationId: getSubscriptionStatus
      summary: 取得當前使用者的訂閱狀態
      description: 依 email（query 或 Bearer）檢查是否有任一有效訂閱。
      parameters:
        - name: email
          in: query
          required: false
          schema: { type: string }
          description: 訂閱時使用的 Email（與 Bearer 二選一）
      responses:
        '200':
          description: 訂閱狀態
          content:
            application/json:
              schema:
                type: object
                properties:
                  hasSubscription: { type: boolean, description: 是否有任一有效訂閱 }

  /reading-record/last-day:
    get:
      operationId: getLastReadingDay
      summary: 取得使用者在指定手冊的最後閱讀天數
      description: 依 email（query 或 Bearer）與 book_id 從閱讀紀錄找出該手冊最後閱讀天數，並回傳 next_day。
      parameters:
        - name: email
          in: query
          required: false
          schema: { type: string }
          description: 訂閱時使用的 Email（與 Bearer 二選一）
        - name: book_id
          in: query
          required: true
          schema: { type: string }
          description: 手冊 ID（從 /books 或 /email/books 取得）
      responses:
        '200':
          description: 最後閱讀天數與下一天
          content:
            application/json:
              schema:
                type: object
                properties:
                  book_id: { type: string }
                  last_day: { type: integer, nullable: true, description: 該書最後閱讀天數；無紀錄則為 null }
                  next_day: { type: integer, description: 建議讀取天數（last_day+1；無紀錄則為 1） }